--- piet-gpu-hal/examples/metal_toy.rs
@@ -1,25 +1,0 @@
-// Copyright 2021 The piet-gpu authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Also licensed under MIT license, at your choice.
-
-//! An example to exercise the Metal backend. Once that becomes
-//! functional, this file will go away.
-
-use piet_gpu_hal::metal;
-
-fn main() {
-    let instance = metal::MetalInstance;
-    println!("hello metal");
-}
