+++ piet-gpu-hal/examples/shader/build.ninja
@@ -1,10 +1,19 @@
 # Build file for shaders.
 
-# You must have glslangValidator in your path, or patch here.
+# You must have Vulkan tools in your path, or patch here.
 
 glslang_validator = glslangValidator
+spirv_cross = spirv-cross
 
 rule glsl
   command = $glslang_validator -V -o $out $in
 
-build collatz.spv: glsl collatz.comp
+rule hlsl
+  command = $spirv_cross --hlsl $in --output $out
+
+rule msl
+  command = $spirv_cross --msl $in --output $out
+
+build gen/collatz.spv: glsl collatz.comp
+build gen/collatz.hlsl: hlsl gen/collatz.spv
+build gen/collatz.msl: msl gen/collatz.spv
